Retrieve hostname from cluster metadata
Make `MetadataFunctionRetriever` retrieve the hostname from the cluster
metadata rather than from the function metadata.
An Azure Functions upstream is associated with a single Azure Function
App, hence all functions of the upstream share the same hostname. This
modification is intended to make it easier for Gloo to populate the
metadata.